+++ src/main/java/org/example/hansabal/domain/review/dto/response/CreateReviewResponse.java
@@ -12,7 +12,6 @@ public class CreateReviewResponse {
     private final String nickname;
     private final String content;
 
-
     public static CreateReviewResponse from(Review review) {
         return new CreateReviewResponse(review.getId(),review.getUser().getNickname(),review.getContent());
     }
